pyproject.toml
[project]
name = "aliyun-fc-acme"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "acme>=2.0.0",
    "josepy>=1.0.0",
    "cryptography>=42.0.0",
    "dnspython>=2.6.0",
    "alibabacloud-cas20200407>=3.0.0",
    "alibabacloud-alidns20150109>=3.0.0",
    "alibabacloud-tea-openapi>=0.3.0",
    "alibabacloud-tea-util>=0.3.0",
    "oss2>=2.18.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501"]
